requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
sklearnmy